pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-backup"
version = "1.0.0"
description = "Secure, deduplicated backup system with S3 support"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "WisePepper-tech"}]
requires-python = ">=3.13"
dependencies = [
    "cryptography>=42.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "slowapi>=0.1.9",
    "argon2-cffi>=25.1.0"
]

[project.scripts]
smart-backup = "main:main"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:slowapi.*",
]